packages/plugin-core/pyproject.toml
[project]
name = "opentools-plugin-core"
version = "0.1.0"
description = "Plugin marketplace core library for OpenTools"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.28",
    "filelock>=3.16",
    "rich>=13.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
sigstore = ["sigstore>=3.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "hypothesis>=6.100",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opentools_plugin_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
